Use correct variable types for Money/Number type as required by gb dataset
and allow the user to define the precision of the float value. As some db like
MYSQL distinguish FLOAT and DECIMAL as 32/64 respectfully.

Also the MoneyType was float but actually a dataset requires MoneyType to be
an integer and for the user to send it as the smallest domination for that
currency. So MoneyType is now int64 and the NumberType can be of int or float
so by default its an int unless the user specify the float precision.